requirements.txt
SQLAlchemy~=2.0.43
loguru~=0.7.3
fastapi~=0.116.1
aiofiles~=24.1.0
pydantic~=2.12.0a1
uvicorn~=0.35.0
python-dotenv~=1.1.1
pillow~=11.3.0